Moved CodekastenParser to core
