+++ src/core/index.ts
@@ -1,8 +1,10 @@
 import { NoteGraph } from "./NoteGraph";
 import { MarkdownLink } from "./Link";
 import { Note} from "./Note";
+import { CodekastenParser } from "./Parser";
 
 export { 
+    CodekastenParser, 
     MarkdownLink,
     Note,
     NoteGraph
