--- src/vscode/index.ts
@@ -1,5 +1,0 @@
-import { CodekastenParser } from "./CodekastenParser";
-
-export {
-    CodekastenParser
-};
